reorganize mutiple components into `components/` including: - _404.jsx - UserAvatar.jsx - LoadingIndicator.jsx - AvatarList.jsx
Refactor original SnackBar to SnackbarView as container component and Snackbar as presentational component
