+++ assets/app/js/components/Common/AvatarList.jsx
