+++ assets/app/js/components/Common/LoadingIndicator.jsx
