+++ assets/app/js/components/Common/UserAvatar.jsx
