+++ assets/app/js/components/Common/_404.jsx
