+++ assets/app/js/components/MilestoneView.jsx
@@ -16,7 +16,7 @@ import MilestoneModal from './MilestoneModal.jsx'
 import MilestoneRow from './MilestoneRow.jsx'
 import AssigneeRow from './AssigneeRow.jsx'
 import Remove from './../icons/Remove.jsx'
-import AvatarList from './AvatarList.jsx'
+import AvatarList from './Common/AvatarList.jsx'
 
 
 class MilestoneView extends Component {
