+++ assets/app/js/components/Notification/NotificationItem.jsx
@@ -1,6 +1,6 @@
 import React, { PropTypes } from 'react';
 
-import UserAvatar from '../UserAvatar.jsx';
+import UserAvatar from '../Common/UserAvatar.jsx';
 
 const propTypes = {
   text: PropTypes.string.isRequired,
